Move superpowers to admin namespace, Replace superpowers index and show with React, update tests
