+++ app/views/admin/superpowers/index.html.erb
@@ -9,5 +9,5 @@
 </ul>
 
 <%if !current_user.nil? && current_user.admin? %>
-  <%= link_to "Add new superpower", new_superpower_path %>
+  <%= link_to "Add new superpower", new_admin_superpower_path %>
 <% end %>
