+++ app/views/admin/superpowers/new.html.erb
