+++ app/views/homes/index.html.erb
@@ -1,4 +1,2 @@
-<p>
-  Modify me in `app/views/homes/index.html.erb` or change the root directive in
-  `config/routes.rb`
-</p>
+<div id="app"></div>
+<%= javascript_pack_tag 'application' %>
