+++ app/views/layouts/application.html.erb
@@ -4,7 +4,7 @@
   <title>SuperpowerReviewSite</title>
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <%= stylesheet_link_tag 'application', media: 'all' %>
-    <%= javascript_pack_tag 'application' %>
+  
     <%= javascript_include_tag "vendor/modernizr" %>
 
   <%= csrf_meta_tags %>
@@ -35,8 +35,6 @@
       <%= value %>
     </div>
   <%- end -%>
-
-  <div id="app"></div>
 
   <%= yield %>
