+++ app/views/reviews/edit.html.erb
@@ -1,6 +1,6 @@
 <h1> Edit Review </h1>
 
-<%= form_for [@review] do |f| %>
+<%= form_for [@superpower, @review] do |f| %>
 
 <div>
   <%= f.label :rating, 'Rating'%>
