--- app/views/superpowers/show.html.erb
@@ -1,26 +1,0 @@
-<h2><%=@superpower.name%></h2>
-
-<div><%=@superpower.description%></div>
-
-
-<%if !current_user.nil? && current_user.admin? %>
-  <%= link_to 'Edit', edit_superpower_path(@superpower) %>
-  <%= link_to 'Delete', @superpower, method: :delete, data: { confirm: 'Are you sure?' } %>
-<%end%>
-
-<%= link_to 'View all superpowers', superpowers_path %>
-
-<h1>Reviews</h1>
-
-<h3><%= link_to 'Add Review', new_superpower_review_path(@superpower) %></h3>
-<% @reviews.each do |review| %>
-  <div>
-    Rating: <%= review.rating %>
-    <%= review.body %>
-
-    <%if !current_user.nil? && current_user.id == review.user_id%>
-      <%= link_to 'Edit Review', edit_review_path(review) %>
-      <%= link_to 'Delete Review', review_path(review), method: :delete, data: { confirm: 'Are you sure?' } %>
-    <%end%>
-  </div>
-<% end %>
