+++ config/routes.rb
@@ -1,15 +1,13 @@
 Rails.application.routes.draw do
-  root 'homes#index'
   devise_for :users
 
-  resources :superpowers do
-    resources :reviews, only: [:new, :create]
+  resources :superpowers, only: [] do
+    resources :reviews, only: [:new, :create, :edit, :update, :destroy]
   end
-
-  resources :reviews, only: [:edit, :update, :destroy]
 
   namespace :admin do
     resources :users
+    resources :superpowers
   end
 
   namespace :api do
@@ -17,5 +15,8 @@ Rails.application.routes.draw do
       resources :superpowers, only: [:index, :show]
     end
   end
+
+  root 'homes#index'
+  get "*path", to: "homes#index"
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 end
